Copy/Paste works (bugs)
Copy/Paste works for columns and rows within tables and to external
programs. But there is a strange bug. When an empty cell is copied to a
destination cell with contents, the contents remain if the empty cell(s)
in the copied area are at the end of the selected area.